Set tags to be nvim version dependent
Debian currently uses an older version of neovim, but
this will eventually change.

For a few plugins, we need to specify a custom tag for a
version that will work with that version of neovim.

Added code to detect the version of neovim and adjust
the tags if an older version is required.